+++ cypress/support/index.d.ts
@@ -1,5 +1,6 @@
 import "./commands";
 import type { Interception, WaitOptions } from "cypress/types/net-stubbing";
+import "../../support";
 
 declare global {
 	namespace Cypress {
@@ -8,15 +9,6 @@ declare global {
 				url: string,
 				options?: { count?: number; delay?: number },
 			): Chainable<void>;
-			waitOnLast<T = any>(
-				alias: string,
-				validate?: (data: Interception<T>) => boolean,
-				options?: {
-					errorMessage?: string;
-					requireValidation?: boolean;
-					displayMessage?: string;
-				} & Partial<WaitOptions>,
-			): Chainable<Interception<T>>;
 		}
 	}
 }